requirements.txt
requests
beautifulsoup4
python-dotenv
pandas
openai
fake-useragent
playwright
